Timings refactor
