+++ src/main/java/cn/nukkit/timings/LevelTimings.java
@@ -1,6 +1,8 @@
 package cn.nukkit.timings;
 
 import cn.nukkit.level.Level;
+import co.aikar.timings.Timing;
+import co.aikar.timings.TimingsManager;
 
 /**
  * @author Pub4Game
